requirements.txt
wheel==0.40.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
bson==0.5.10
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colorlog==4.8.0
commonmark==0.9.1
dnspython==2.3.0
dotdict==0.1
fastjsonschema==2.16.3
flake8==6.0.0
flake8-polyfill==1.0.2
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
jsonschema==4.17.3
jupyter_core==5.3.0
loguru==0.7.0
magic-filter==1.0.9
mando==0.6.4
mccabe==0.7.0
mongomantic==0.4.2
multidict==6.0.4
nbformat==5.8.0
packaging==23.1
platformdirs==3.5.0
plotly==5.14.1
progress==1.6
pyaes==1.6.1
pycodestyle==2.10.0
pydantic==1.9.2
pyflakes==3.0.1
Pygments==2.15.1
pymodm==0.4.3
pymongo==3.13.0
Pyrogram==2.0.106
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
radon==4.0.0
requests==2.30.0
rich==9.13.0
six==1.16.0
smmap==5.0.0
tabulate==0.9.0
tenacity==8.2.2
TgCrypto==1.2.5
traitlets==5.9.0
typing-extensions==3.10.0.2
urllib3==2.0.2
wily==1.24.0
win32-setctime==1.1.0
yarl==1.9.2
